Refactored direct call bindings to be more explicitly named
