+++ Assets/dotnow/Scripts/Integration/Unity/DirectCallBindings.Transform.cs
@@ -8,40 +8,8 @@ using dotnow.Runtime;
 namespace UnityEngine
 {
 
-    internal static class DirectCallBindings
+    internal static partial class DirectCallBindings
     {
-        [Preserve]
-        [CLRMethodDirectCallBinding(typeof(Time), "get_time")]
-        public static void UnityEngine_Time_GetTime(StackData[] stack, int offset)
-        {
-            stack[offset].value.Single = Time.time;
-            stack[offset].type = StackData.ObjectType.Single;
-        }
-
-        [Preserve]
-        [CLRMethodDirectCallBinding(typeof(Input), "GetKey", typeof(KeyCode))]
-        public static void UnityEngine_Input_GetKey(StackData[] stack, int offset)
-        {
-            stack[offset].value.Int32 = Input.GetKey((KeyCode)stack[offset].value.Int32) ? 1 : 0;
-            stack[offset].type = StackData.ObjectType.Int32;
-        }
-
-        [Preserve]
-        [CLRMethodDirectCallBinding(typeof(GameObject), "get_transform")]
-        public static void unityEngine_GameObject_GetTransform(StackData[] stack, int offset)
-        {
-            stack[offset].refValue = ((GameObject)stack[offset].refValue).transform;
-            stack[offset].type = StackData.ObjectType.Ref;
-        }
-
-        [Preserve]
-        [CLRMethodDirectCallBinding(typeof(Component), "get_transform")]
-        public static void unityEngine_Component_GetTransform(StackData[] stack, int offset)
-        {
-            stack[offset].refValue = ((Component)stack[offset].refValue.Unwrap()).transform;
-            stack[offset].type = StackData.ObjectType.Ref;
-        }
-
         [Preserve]
         [CLRMethodDirectCallBinding(typeof(Transform), "Rotate", typeof(float), typeof(float), typeof(float))]
         public static void UnityEngine_Transform_Rotate_SingleSingleSingle(StackData[] stack, int offset)
